+++ steel_vm/src/contracts.rs
@@ -1,12 +1,12 @@
 use crate::{evaluation_progress::EvaluationProgress, heap::Heap, vm::vm};
 use std::{cell::RefCell, rc::Rc};
 use steel::{
+    compiler::constants::ConstantTable,
     env::Env,
     gc::Gc,
     parser::span::Span,
     rerrs::{ErrorKind, SteelErr},
     rvals::{ByteCodeLambda, Result, SteelVal},
-    steel_compiler::constants::ConstantTable,
     stop,
     values::contracts::{ContractType, ContractedFunction, FlatContract, FunctionContract},
 };
